+++ CS345 Pente/CS345 Pente/CSPPenteGameModelInterface.h
@@ -25,7 +25,7 @@ typedef enum {
 -(void)makeMove:(id<CSPMoveInterface>) aMove;
 
 // How many stones has player ___ captured? (0 through 10)
--(int) capturesByPlayer:(CSPPlayerID) player;
+-(NSUInteger) capturesByPlayer:(CSPPlayerID) player;
 
 // Who has won the game? (White/Black/Nobody yet)
 -(CSPPlayerID) whoIsWinner;
@@ -37,12 +37,12 @@ typedef enum {
 -(CSPPlayerID) whosePieceIsAt:(id<CSPCoordinateInterface>) aLocation;
 
 // How wide is the board?
--(int) boardWidth;
+-(NSUInteger) boardWidth;
 
 
 // Low priority:
 // Max adjacent stones of stone ___? (0-4)
 // [For user-assistance coloring purposes, still allowing sneaky XX-XX wins.]
--(int)longestChainForStoneAt:(id<CSPCoordinateInterface>) aLocation;
+-(NSUInteger)longestChainForStoneAt:(id<CSPCoordinateInterface>) aLocation;
 
 @end
