+++ CS345 Pente/CS345 Pente/CSPStone.h
@@ -16,12 +16,12 @@
 -(id) initWithPlacement: (id<CSPMoveInterface>) whoAndWhere;
 -(void) setNeighborTo:(CSPStone*) newNeighbor
           inDirection:(CSPDirectionID) whichWay;
--(int) chainLengthIn:(CSPDirectionID) whichDirection;
+-(NSUInteger) chainLengthIn:(CSPDirectionID) whichDirection;
 -(NSSet*) bookendedStonesIn:(CSPDirectionID) whichDirection;
 -(NSSet*) bookendedStones;
--(int) neighborCount;
+-(NSUInteger) neighborCount;
 -(CSPStone*) neighborInDirection:(CSPDirectionID) whichDirection;
--(int) longestChainLength;
+-(NSUInteger) longestChainLength;
 @end
 
 CSPDirectionID oppositeDirectionOf(CSPDirectionID aDirection);
